FEATURE: rename MainFeedPage.tsx to MainFeed.tsx, and update main.tsx based on rename
